Refine local-server step in `mlflow agent setup`
- Move the local-server section so it runs after `mlflow` is installed,
  and give the agent the literal `mlflow server` command with the
  CLI-picked port baked in. The server is left running after
  verification so the user can open the trace URL in the UI.
- Split `build_task` out of `skill_installer.py` into a dedicated
  `task.py` module.
